+++ lib/maybe_just_nothing.dart
@@ -1,12 +1,15 @@
 import 'dart:async';
 
 /// A variation of the Maybe monad with eager execution.
-abstract class Maybe<T extends Object?> {
+abstract class Maybe<T> {
   /// Maps the value to P.
-  Maybe<P> map<P extends Object?>(P Function(T value) mapper);
+  Maybe<P> map<P>(P Function(T value) mapper);
+
+  /// Tries to map value to P. On exception produces Nothing.
+  Maybe<P> tryMap<P>(P Function(T value) mapper);
 
   /// Maps the value to P.
-  Maybe<P> flatMap<P extends Object?>(Maybe<P> Function(T value) mapper);
+  Maybe<P> flatMap<P>(Maybe<P> Function(T value) mapper);
 
   /// Filter the value using the [predicate].
   Maybe<T> where(bool Function(T value) predicate);
@@ -33,12 +36,11 @@ abstract class Maybe<T extends Object?> {
   void ifNothing(void Function() callback);
 
   /// Narrows the type to P if the value is present and has actually the type of P.
-  Maybe<P> type<P extends Object?>();
+  Maybe<P> type<P>();
 
   /// If this and the other are both [Just] values, merges them using the [merger] function and returns [Just]<V>.
   /// Otherwise returns [Nothing]<V>
-  Maybe<R> merge<R extends Object, V extends Object?>(
-      Maybe<V> other, Merger<R, T, V> merger);
+  Maybe<R> merge<R, V>(Maybe<V> other, Merger<R, T, V> merger);
 
   /// If this is [Nothing], returns the result of [next]. Otherwise return this..
   Maybe<T> fallback(Maybe<T> Function() next);
@@ -48,19 +50,26 @@ abstract class Maybe<T extends Object?> {
 }
 
 /// Represents an existing value of type T.
-class Just<T extends Object?> implements Maybe<T> {
-  Just(this.value);
+class Just<T> implements Maybe<T> {
+  const Just(this.value);
 
   /// The wrapped value.
   final T value;
 
   @override
-  Maybe<P> map<P extends Object?>(P Function(T value) mapper) =>
-      Just<P>(mapper(value));
+  Maybe<P> map<P>(P Function(T value) mapper) => Just(mapper(value));
 
   @override
-  Maybe<P> flatMap<P extends Object?>(Maybe<P> Function(T value) mapper) =>
-      mapper(value);
+  Maybe<P> tryMap<P>(P Function(T value) mapper) {
+    try {
+      return Just(mapper(value));
+    } catch (e) {
+      return Nothing();
+    }
+  }
+
+  @override
+  Maybe<P> flatMap<P>(Maybe<P> Function(T value) mapper) => mapper(value);
 
   @override
   T or(T defaultValue) => value;
@@ -85,15 +94,13 @@ class Just<T extends Object?> implements Maybe<T> {
 
   @override
   Maybe<T> where(bool Function(T value) predicate) =>
-      predicate(value) ? this : Nothing<T>();
+      predicate(value) ? this : Nothing();
 
   @override
-  Maybe<P> type<P extends Object?>() =>
-      value is P ? Just(value as P) : Nothing<P>();
+  Maybe<P> type<P>() => value is P ? Just(value as P) : Nothing();
 
   @override
-  Maybe<R> merge<R extends Object, V extends Object?>(
-          Maybe<V> other, Merger<R, T, V> merger) =>
+  Maybe<R> merge<R, V>(Maybe<V> other, Merger<R, T, V> merger) =>
       flatMap((a) => other.map((b) => merger(a, b)));
 
   @override
@@ -112,15 +119,17 @@ class Just<T extends Object?> implements Maybe<T> {
 typedef Merger<R, T, V> = R Function(T a, V b);
 
 /// Represents a non-existing value of type T.
-class Nothing<T extends Object?> implements Maybe<T> {
+class Nothing<T> implements Maybe<T> {
   const Nothing();
 
   @override
-  Nothing<P> map<P extends Object?>(P Function(T value) mapper) => Nothing<P>();
+  Nothing<P> map<P>(P Function(T value) mapper) => Nothing();
 
   @override
-  Nothing<P> flatMap<P extends Object?>(Maybe<P> Function(T value) mapper) =>
-      Nothing<P>();
+  Nothing<P> tryMap<P>(P Function(T value) mapper) => Nothing();
+
+  @override
+  Nothing<P> flatMap<P>(Maybe<P> Function(T value) mapper) => Nothing();
 
   @override
   T or(T defaultValue) => defaultValue;
@@ -147,12 +156,10 @@ class Nothing<T extends Object?> implements Maybe<T> {
   Nothing<T> where(bool Function(T value) predicate) => this;
 
   @override
-  Nothing<P> type<P extends Object?>() => Nothing<P>();
+  Nothing<P> type<P>() => Nothing();
 
   @override
-  Nothing<R> merge<R extends Object, V extends Object?>(
-          Maybe<V> other, Merger<R, T, V> merger) =>
-      Nothing<R>();
+  Nothing<R> merge<R, V>(Maybe<V> other, Merger<R, T, V> merger) => Nothing();
 
   @override
   Maybe<T> fallback(Maybe<T> Function() next) => next();